.github/release/pyproject.toml
[project]
name = "release"
version = "0.0.1"
requires-python = ">=3.13,<4"
dependencies = [
    "semver (>=3.0.4,<4.0.0)",
    "pygithub (>=2.6.1,<3.0.0)",
    "gitpython (>=3.1.44,<4.0.0)"
]

[project.scripts]
release = "release:main"

[build-system]
requires = ["poetry-core>=2.1.2,<3.0.0"]
build-backend = "poetry.core.masonry.api"
